[REFACTOR] 디렉토리 분리 - 어르신 매칭 상태 조회 (matchingStatus)
