+++ src/main/java/com/balybus/galaxy/patient/controller/PatientController.java
@@ -194,45 +194,45 @@ public class PatientController {
 
     ////
 
-    @Operation(summary = "어르신 매칭 대기 중 리스트 반환 API", description = "매칭 대기 중인 어르신 정보를 반환합니다.")
-    @ApiResponses({
-            @ApiResponse(responseCode = "200", description = "반환 성공",
-                    content = @Content(schema = @Schema(implementation = PatientResponseDto.UpdateRecruitPatientInfo.class))),
-            @ApiResponse(responseCode = "4008", description = "사용자정의에러코드:로그인이 정보 없음(쿠키 없음)",
-                    content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
-            @ApiResponse(responseCode = "7000", description = "사용자정의에러코드:로그인한 사용자의 권한이 매니저가 아닙니다.",
-                    content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
-    })
-    @GetMapping("/get-matching-waiting-patient-list")
-    public ResponseEntity<?> getMatchingWaitingPatientList(@AuthenticationPrincipal UserDetails userDetails) {
-        return ResponseEntity.ok().body(patientService.matchingPatientInfoList(userDetails.getUsername()));
-    }
-
-    @Operation(summary = "어르신 매칭 중 리스트 반환 API", description = "매칭 중인 어르신 정보를 반환합니다.")
-    @ApiResponses({
-            @ApiResponse(responseCode = "200", description = "반환 성공",
-                    content = @Content(schema = @Schema(implementation = PatientResponseDto.UpdateRecruitPatientInfo.class))),
-            @ApiResponse(responseCode = "4008", description = "사용자정의에러코드:로그인이 정보 없음(쿠키 없음)",
-                    content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
-            @ApiResponse(responseCode = "7000", description = "사용자정의에러코드:로그인한 사용자의 권한이 매니저가 아닙니다.",
-                    content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
-    })
-    @GetMapping("/get-matching-patient-list")
-    public ResponseEntity<?> getMatchingPatientList(@AuthenticationPrincipal UserDetails userDetails) {
-        return ResponseEntity.ok().body(patientService.matchingPatientInfoList(userDetails.getUsername()));
-    }
-
-    @Operation(summary = "어르신 매칭 완료 리스트 반환 API", description = "매칭 완료 / 거절 상태인 어르신 정보를 반환합니다.")
-    @ApiResponses({
-            @ApiResponse(responseCode = "200", description = "반환 성공",
-                    content = @Content(schema = @Schema(implementation = PatientResponseDto.UpdateRecruitPatientInfo.class))),
-            @ApiResponse(responseCode = "4008", description = "사용자정의에러코드:로그인이 정보 없음(쿠키 없음)",
-                    content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
-            @ApiResponse(responseCode = "7000", description = "사용자정의에러코드:로그인한 사용자의 권한이 매니저가 아닙니다.",
-                    content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
-    })
-    @GetMapping("/get-matched-patient-list")
-    public ResponseEntity<?> getMatchedPatientList(@AuthenticationPrincipal UserDetails userDetails) {
-        return ResponseEntity.ok().body(patientService.matchedPatientInfoList(userDetails.getUsername()));
-    }
+//    @Operation(summary = "어르신 매칭 대기 중 리스트 반환 API", description = "매칭 대기 중인 어르신 정보를 반환합니다.")
+//    @ApiResponses({
+//            @ApiResponse(responseCode = "200", description = "반환 성공",
+//                    content = @Content(schema = @Schema(implementation = PatientResponseDto.UpdateRecruitPatientInfo.class))),
+//            @ApiResponse(responseCode = "4008", description = "사용자정의에러코드:로그인이 정보 없음(쿠키 없음)",
+//                    content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
+//            @ApiResponse(responseCode = "7000", description = "사용자정의에러코드:로그인한 사용자의 권한이 매니저가 아닙니다.",
+//                    content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
+//    })
+//    @GetMapping("/get-matching-waiting-patient-list")
+//    public ResponseEntity<?> getMatchingWaitingPatientList(@AuthenticationPrincipal UserDetails userDetails) {
+//        return ResponseEntity.ok().body(patientService.matchingPatientInfoList(userDetails.getUsername()));
+//    }
+//
+//    @Operation(summary = "어르신 매칭 중 리스트 반환 API", description = "매칭 중인 어르신 정보를 반환합니다.")
+//    @ApiResponses({
+//            @ApiResponse(responseCode = "200", description = "반환 성공",
+//                    content = @Content(schema = @Schema(implementation = PatientResponseDto.UpdateRecruitPatientInfo.class))),
+//            @ApiResponse(responseCode = "4008", description = "사용자정의에러코드:로그인이 정보 없음(쿠키 없음)",
+//                    content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
+//            @ApiResponse(responseCode = "7000", description = "사용자정의에러코드:로그인한 사용자의 권한이 매니저가 아닙니다.",
+//                    content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
+//    })
+//    @GetMapping("/get-matching-patient-list")
+//    public ResponseEntity<?> getMatchingPatientList(@AuthenticationPrincipal UserDetails userDetails) {
+//        return ResponseEntity.ok().body(patientService.matchingPatientInfoList(userDetails.getUsername()));
+//    }
+//
+//    @Operation(summary = "어르신 매칭 완료 리스트 반환 API", description = "매칭 완료 / 거절 상태인 어르신 정보를 반환합니다.")
+//    @ApiResponses({
+//            @ApiResponse(responseCode = "200", description = "반환 성공",
+//                    content = @Content(schema = @Schema(implementation = PatientResponseDto.UpdateRecruitPatientInfo.class))),
+//            @ApiResponse(responseCode = "4008", description = "사용자정의에러코드:로그인이 정보 없음(쿠키 없음)",
+//                    content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
+//            @ApiResponse(responseCode = "7000", description = "사용자정의에러코드:로그인한 사용자의 권한이 매니저가 아닙니다.",
+//                    content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
+//    })
+//    @GetMapping("/get-matched-patient-list")
+//    public ResponseEntity<?> getMatchedPatientList(@AuthenticationPrincipal UserDetails userDetails) {
+//        return ResponseEntity.ok().body(patientService.matchedPatientInfoList(userDetails.getUsername()));
+//    }
 }
