+++ src/main/java/com/balybus/galaxy/patient/service/PatientService.java
@@ -20,6 +20,6 @@ public interface PatientService {
 
 
 
-    PatientResponseDto.MatchingPatientInfoList matchingPatientInfoList(String userEmail); // 어르신 매칭중 리스트 반화
-    PatientResponseDto.MatchedPatientInfoList matchedPatientInfoList(String userEmail); // 어르신 매칭 완료 리스트 반화
+//    PatientResponseDto.MatchingPatientInfoList matchingPatientInfoList(String userEmail); // 어르신 매칭중 리스트 반화
+//    PatientResponseDto.MatchedPatientInfoList matchedPatientInfoList(String userEmail); // 어르신 매칭 완료 리스트 반화
 }
