+++ src/main/java/com/balybus/galaxy/patient/service/serviceImpl/PatientServiceImpl.java
@@ -522,126 +522,126 @@ public class PatientServiceImpl implements PatientService {
 //                .build();
 //    }
 
-    /**
-     * 매칭중 어르신 정보 반환(관리자 입장)
-     * @param userEmail 관리자 이메일
-     * @return
-     */
-    @Override
-    public PatientResponseDto.MatchingPatientInfoList matchingPatientInfoList(String userEmail) {
-        // 1. 관리자 정보 조회
-        TblCenterManager centerManager = loginAuthCheckService.checkManager(userEmail);
-
-        // 2. 연관된 어르신 정보 로그 테이블 모두 조회
-        List<TblPatientLog> tblPatientLog = patientLogRepository.findAllByManagerId(centerManager.getId());
-
-        // 매칭중인 어르신 리스트(최종 반환 값)
-        List<PatientResponseDto.MatchingPatientInfo> matchingPatientInfoList = new ArrayList<>();
-
-        for(TblPatientLog ptLog : tblPatientLog) {
-            // 1. 어르신 매칭 테이블 반환(매칭 중 상태이며 해당 어르신 관계에 있는 테이블)
-            List<TblMatching> tblMatching = tblMatchingRepository.findByPatientLog_idAndMatchState(ptLog.getId(), INIT);
-
-            // 2. 매칭 중인 어르신의 요양 보호사 리스트
-            List<PatientResponseDto.MatchedHelperInfo> matchedHelperInfoList = new ArrayList<>();
-
-            // 3. 어르신 매칭 상태별 해당하는 요양 보호사 반환
-            for(TblMatching matching : tblMatching) {
-                // 4. 요양 보호사 검색
-                TblHelper tblHelper = helperRepository.findById(matching.getHelper().getId())
-                        .orElseThrow(() -> new BadRequestException(NOT_FOUND_HELPER));
-
-                matchedHelperInfoList.add(PatientResponseDto.MatchedHelperInfo.builder()
-                                .helperSeq(tblHelper.getId())
-                                .name(tblHelper.getName())
-                                .gender(tblHelper.getGender())
-                                .age(tblHelper.getBirthday())
-                        .build());
-            }
-            // 5. 매칭 중인 어르신 정보
-            PatientResponseDto.MatchingPatientInfo matchingPatientInfo = PatientResponseDto.MatchingPatientInfo.builder()
-                    .patientSeq(ptLog.getPatient().getId())
-                    .name(ptLog.getName())
-                    .gender(ptLog.getGender())
-                    .birthDate(ptLog.getBirthDate())
-                    .workType(ptLog.getPatient().getWorkType())
-                    .tblAddressFirst(ptLog.getTblAddressFirst().getName())
-                    .tblAddressSecond(ptLog.getTblAddressSecond().getName())
-                    .tblAddressThird(ptLog.getTblAddressThird().getName())
-                    .matchedHelperInfos(matchedHelperInfoList)
-                    .build();
-            matchingPatientInfoList.add(matchingPatientInfo);
-        }
-
-        return PatientResponseDto.MatchingPatientInfoList.builder()
-                .matchingPatientInfoList(matchingPatientInfoList)
-                .build();
-    }
-
-    /**
-     * 매칭 완료(수락) 및 매칭 거절 상태인 어르신 정보 리스트 반환(관리자 입장)
-     * @param userEmail 관리자 이메일
-     * @return
-     */
-    @Override
-    public PatientResponseDto.MatchedPatientInfoList matchedPatientInfoList(String userEmail) {
-        // 1. 관리자 정보 조회
-        TblCenterManager centerManager = loginAuthCheckService.checkManager(userEmail);
-
-        // 2. 연관된 어르신 정보 로그 테이블 모두 조회
-        List<TblPatientLog> tblPatientLog = patientLogRepository.findAllByManagerId(centerManager.getId());
-
-        // 매칭 완료(수락) 상태인 어르신 리스트(최종 반환 값)
-        List<PatientResponseDto.MatchingPatientInfo> matchingPatientInfoList = getPatientListByMatchState(tblPatientLog, MATCH_FIN);
-
-        // 매칭 거절 상태인 어르신 리스트(최종 반환 값)
-        List<PatientResponseDto.MatchingPatientInfo> matchingRejectedHelperInfoList = getPatientListByMatchState(tblPatientLog, REJECT);
-
-        return PatientResponseDto.MatchedPatientInfoList.builder()
-                .matchedPatientInfoList(matchingPatientInfoList)
-                .matchingRejectedPatientInfoList(matchingRejectedHelperInfoList)
-                .build();
-    }
-
-    private List<PatientResponseDto.MatchingPatientInfo> getPatientListByMatchState(List<TblPatientLog> tblPatientLog, MatchState matchState) {
-        List<PatientResponseDto.MatchingPatientInfo> matchingPatientInfoList = new ArrayList<>();
-
-        for(TblPatientLog ptLog : tblPatientLog) {
-            // 1. 어르신 매칭 테이블 반환(매칭 완료 상태이며 해당 어르신 관계에 있는 테이블)
-            List<TblMatching> tblMatching = tblMatchingRepository.findByPatientLog_idAndMatchState(ptLog.getId(), matchState);
-
-            // 2. 어르신의 요양 보호사 리스트
-            List<PatientResponseDto.MatchedHelperInfo> matchedHelperInfoList = new ArrayList<>();
-
-            // 3. 어르신 매칭 상태별 해당하는 요양 보호사 반환
-            for(TblMatching matching : tblMatching) {
-                // 4. 요양 보호사 검색
-                TblHelper tblHelper = helperRepository.findById(matching.getHelper().getId())
-                        .orElseThrow(() -> new BadRequestException(NOT_FOUND_HELPER));
-                matchedHelperInfoList.add(PatientResponseDto.MatchedHelperInfo.builder()
-                        .helperSeq(tblHelper.getId())
-                        .name(tblHelper.getName())
-                        .gender(tblHelper.getGender())
-                        .age(tblHelper.getBirthday())
-                        .build());
-            }
-            // 5. 매칭 중인 어르신 정보
-            if(!matchedHelperInfoList.isEmpty()) {
-                PatientResponseDto.MatchingPatientInfo matchingPatientInfo = PatientResponseDto.MatchingPatientInfo.builder()
-                        .patientSeq(ptLog.getPatient().getId())
-                        .name(ptLog.getName())
-                        .gender(ptLog.getGender())
-                        .birthDate(ptLog.getBirthDate())
-                        .workType(ptLog.getPatient().getWorkType())
-                        .tblAddressFirst(ptLog.getTblAddressFirst().getName())
-                        .tblAddressSecond(ptLog.getTblAddressSecond().getName())
-                        .tblAddressThird(ptLog.getTblAddressThird().getName())
-                        .matchedHelperInfos(matchedHelperInfoList)
-                        .build();
-                matchingPatientInfoList.add(matchingPatientInfo);
-            }
-        }
-        return matchingPatientInfoList;
-    }
+//    /**
+//     * 매칭중 어르신 정보 반환(관리자 입장)
+//     * @param userEmail 관리자 이메일
+//     * @return
+//     */
+//    @Override
+//    public PatientResponseDto.MatchingPatientInfoList matchingPatientInfoList(String userEmail) {
+//        // 1. 관리자 정보 조회
+//        TblCenterManager centerManager = loginAuthCheckService.checkManager(userEmail);
+//
+//        // 2. 연관된 어르신 정보 로그 테이블 모두 조회
+//        List<TblPatientLog> tblPatientLog = patientLogRepository.findAllByManagerId(centerManager.getId());
+//
+//        // 매칭중인 어르신 리스트(최종 반환 값)
+//        List<PatientResponseDto.MatchingPatientInfo> matchingPatientInfoList = new ArrayList<>();
+//
+//        for(TblPatientLog ptLog : tblPatientLog) {
+//            // 1. 어르신 매칭 테이블 반환(매칭 중 상태이며 해당 어르신 관계에 있는 테이블)
+//            List<TblMatching> tblMatching = tblMatchingRepository.findByPatientLog_idAndMatchState(ptLog.getId(), INIT);
+//
+//            // 2. 매칭 중인 어르신의 요양 보호사 리스트
+//            List<PatientResponseDto.MatchedHelperInfo> matchedHelperInfoList = new ArrayList<>();
+//
+//            // 3. 어르신 매칭 상태별 해당하는 요양 보호사 반환
+//            for(TblMatching matching : tblMatching) {
+//                // 4. 요양 보호사 검색
+//                TblHelper tblHelper = helperRepository.findById(matching.getHelper().getId())
+//                        .orElseThrow(() -> new BadRequestException(NOT_FOUND_HELPER));
+//
+//                matchedHelperInfoList.add(PatientResponseDto.MatchedHelperInfo.builder()
+//                                .helperSeq(tblHelper.getId())
+//                                .name(tblHelper.getName())
+//                                .gender(tblHelper.getGender())
+//                                .age(tblHelper.getBirthday())
+//                        .build());
+//            }
+//            // 5. 매칭 중인 어르신 정보
+//            PatientResponseDto.MatchingPatientInfo matchingPatientInfo = PatientResponseDto.MatchingPatientInfo.builder()
+//                    .patientSeq(ptLog.getPatient().getId())
+//                    .name(ptLog.getName())
+//                    .gender(ptLog.getGender())
+//                    .birthDate(ptLog.getBirthDate())
+//                    .workType(ptLog.getPatient().getWorkType())
+//                    .tblAddressFirst(ptLog.getTblAddressFirst().getName())
+//                    .tblAddressSecond(ptLog.getTblAddressSecond().getName())
+//                    .tblAddressThird(ptLog.getTblAddressThird().getName())
+//                    .matchedHelperInfos(matchedHelperInfoList)
+//                    .build();
+//            matchingPatientInfoList.add(matchingPatientInfo);
+//        }
+//
+//        return PatientResponseDto.MatchingPatientInfoList.builder()
+//                .matchingPatientInfoList(matchingPatientInfoList)
+//                .build();
+//    }
+//
+//    /**
+//     * 매칭 완료(수락) 및 매칭 거절 상태인 어르신 정보 리스트 반환(관리자 입장)
+//     * @param userEmail 관리자 이메일
+//     * @return
+//     */
+//    @Override
+//    public PatientResponseDto.MatchedPatientInfoList matchedPatientInfoList(String userEmail) {
+//        // 1. 관리자 정보 조회
+//        TblCenterManager centerManager = loginAuthCheckService.checkManager(userEmail);
+//
+//        // 2. 연관된 어르신 정보 로그 테이블 모두 조회
+//        List<TblPatientLog> tblPatientLog = patientLogRepository.findAllByManagerId(centerManager.getId());
+//
+//        // 매칭 완료(수락) 상태인 어르신 리스트(최종 반환 값)
+//        List<PatientResponseDto.MatchingPatientInfo> matchingPatientInfoList = getPatientListByMatchState(tblPatientLog, MATCH_FIN);
+//
+//        // 매칭 거절 상태인 어르신 리스트(최종 반환 값)
+//        List<PatientResponseDto.MatchingPatientInfo> matchingRejectedHelperInfoList = getPatientListByMatchState(tblPatientLog, REJECT);
+//
+//        return PatientResponseDto.MatchedPatientInfoList.builder()
+//                .matchedPatientInfoList(matchingPatientInfoList)
+//                .matchingRejectedPatientInfoList(matchingRejectedHelperInfoList)
+//                .build();
+//    }
+//
+//    private List<PatientResponseDto.MatchingPatientInfo> getPatientListByMatchState(List<TblPatientLog> tblPatientLog, MatchState matchState) {
+//        List<PatientResponseDto.MatchingPatientInfo> matchingPatientInfoList = new ArrayList<>();
+//
+//        for(TblPatientLog ptLog : tblPatientLog) {
+//            // 1. 어르신 매칭 테이블 반환(매칭 완료 상태이며 해당 어르신 관계에 있는 테이블)
+//            List<TblMatching> tblMatching = tblMatchingRepository.findByPatientLog_idAndMatchState(ptLog.getId(), matchState);
+//
+//            // 2. 어르신의 요양 보호사 리스트
+//            List<PatientResponseDto.MatchedHelperInfo> matchedHelperInfoList = new ArrayList<>();
+//
+//            // 3. 어르신 매칭 상태별 해당하는 요양 보호사 반환
+//            for(TblMatching matching : tblMatching) {
+//                // 4. 요양 보호사 검색
+//                TblHelper tblHelper = helperRepository.findById(matching.getHelper().getId())
+//                        .orElseThrow(() -> new BadRequestException(NOT_FOUND_HELPER));
+//                matchedHelperInfoList.add(PatientResponseDto.MatchedHelperInfo.builder()
+//                        .helperSeq(tblHelper.getId())
+//                        .name(tblHelper.getName())
+//                        .gender(tblHelper.getGender())
+//                        .age(tblHelper.getBirthday())
+//                        .build());
+//            }
+//            // 5. 매칭 중인 어르신 정보
+//            if(!matchedHelperInfoList.isEmpty()) {
+//                PatientResponseDto.MatchingPatientInfo matchingPatientInfo = PatientResponseDto.MatchingPatientInfo.builder()
+//                        .patientSeq(ptLog.getPatient().getId())
+//                        .name(ptLog.getName())
+//                        .gender(ptLog.getGender())
+//                        .birthDate(ptLog.getBirthDate())
+//                        .workType(ptLog.getPatient().getWorkType())
+//                        .tblAddressFirst(ptLog.getTblAddressFirst().getName())
+//                        .tblAddressSecond(ptLog.getTblAddressSecond().getName())
+//                        .tblAddressThird(ptLog.getTblAddressThird().getName())
+//                        .matchedHelperInfos(matchedHelperInfoList)
+//                        .build();
+//                matchingPatientInfoList.add(matchingPatientInfo);
+//            }
+//        }
+//        return matchingPatientInfoList;
+//    }
 
 }
